refactor EventsController by moving some methods to Utils namespace
