+++ app/views/partials/attendees/_card.html.erb
@@ -1,3 +1,3 @@
-<turbo-frame id="<%= dom_id attendee %>" class='w-full'>
+<div id="<%= dom_id attendee %>_attendee_card" class='w-full'>
   <%= content_tag(:span, attendee.username, class: 'block') %>
-</turbo-frame>
+</div>
